move authorization tests from TWLight.users to TWLight because they are so cross-cutting.
